+++ package.json
@@ -33,6 +33,7 @@
     "babel-loader": "^7.1.4",
     "classnames": "^2.2.5",
     "express": "^4.16.3",
+    "history": "^4.7.2",
     "html-webpack-plugin": "^3.1.0",
     "ip": "^1.1.5",
     "json-loader": "^0.5.7",
@@ -42,6 +43,8 @@
     "react-dom": "^16.3.0",
     "react-hot-loader": "^4.0.1",
     "react-redux": "^5.0.7",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^4.0.8",
     "redux": "^3.7.2",
     "rimraf": "^2.6.2",
     "uglifyjs-webpack-plugin": "^1.2.4",
