+++ src/app/app.js
@@ -1,0 +1,21 @@
+import { hot } from 'react-hot-loader';
+import { Route, Switch } from 'react-router-dom';
+import React, { Component } from 'react';
+import Addition from '../games/math/addition/addition';
+
+import style from './app.css';
+
+export class App extends Component {
+  render() {
+    return (
+      <div className={style.app}>
+        <h1>Play for reward</h1>
+        <Switch>
+          <Route path="/game" component={Addition} />
+        </Switch>
+      </div>
+    );
+  }
+}
+
+export default hot(module)(App);
