+++ src/root.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { ConnectedRouter } from 'react-router-redux';
+import { Provider } from 'react-redux';
+import App from './app/app';
+import createBrowserHistory from 'history/createBrowserHistory';
+import createStore from './store/createStore';
+
+const initialState = {};
+const history = createBrowserHistory();
+
+const store = createStore(initialState, history);
+
+ReactDOM.render(
+  <Provider store={store}>
+    <ConnectedRouter history={history}>
+      <App />
+    </ConnectedRouter>
+  </Provider>,
+  document.getElementById('root'),
+);
